Shut up stupid warning.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@7253 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/string/strformattype.cpp
@@ -290,7 +290,7 @@ RTDECL(int) RTStrFormatTypeRegister(const char *pszType, PFNRTSTRFORMATTYPE pfnH
             /* insert the new entry. */
             memset(&g_aTypes[i], 0, sizeof(g_aTypes[i]));
             memcpy(&g_aTypes[i].szType[0], pszType, cchType + 1);
-            g_aTypes[i].cchType = cchType;
+            g_aTypes[i].cchType = (uint8_t)cchType;
             g_aTypes[i].pvUser = pvUser;
 #ifdef IN_GC
             g_aTypes[i].offHandler = (intptr_t)pfnHandler - (intptr_t)&g_aTypes[0];
